Find and Remove Edge Loops and Duplicates

This function finds and removes loops, edge duplicates, and bi-directional edges.

edge.duplicates(network, verbose = TRUE)

Arguments

network
An edge list, which is either a data frame or a character or integer matrix with two columns.

verbose
Logical, whether to display the progress of the function on the screen. Defaults to TRUE.

Details

This function removes loops, duplicate and bi-directional edges; the edges that occur closer to the end of the edge list will be removed.

Value

A list with the following elements: edges - a character matrix of the edges in the network with any loops or duplicate edges removed; inds - an integer vector of the edge indices of any loop or duplicate edges in the original network.

References

Kalinka, A.T. and Tomancak, P. (2011). linkcomm: an R package for the generation, visualization, and analysis of link communities in networks of arbitrary size and type. Bioinformatics 27, 2011-2012.

Author

Alex T. Kalinka alex.t.kalinka@gmail.com

Examples

## Generate graph and remove loops and duplicate edges.
g <- swiss[,3:4]
g.dup <- edge.duplicates(g)

#>    Checking for loops and duplicate edges... 0.000%   Checking for loops and duplicate edges... 2.174%   Checking for loops and duplicate edges... 4.348%   Checking for loops and duplicate edges... 6.522%   Checking for loops and duplicate edges... 8.696%   Checking for loops and duplicate edges... 10.870%   Checking for loops and duplicate edges... 13.043%   Checking for loops and duplicate edges... 15.217%   Checking for loops and duplicate edges... 17.391%   Checking for loops and duplicate edges... 19.565%   Checking for loops and duplicate edges... 21.739%   Checking for loops and duplicate edges... 23.913%   Checking for loops and duplicate edges... 26.087%   Checking for loops and duplicate edges... 28.261%   Checking for loops and duplicate edges... 30.435%   Checking for loops and duplicate edges... 32.609%   Checking for loops and duplicate edges... 34.783%   Checking for loops and duplicate edges... 36.957%   Checking for loops and duplicate edges... 39.130%   Checking for loops and duplicate edges... 41.304%   Checking for loops and duplicate edges... 43.478%   Checking for loops and duplicate edges... 45.652%   Checking for loops and duplicate edges... 47.826%   Checking for loops and duplicate edges... 50.000%   Checking for loops and duplicate edges... 52.174%   Checking for loops and duplicate edges... 54.348%   Checking for loops and duplicate edges... 56.522%   Checking for loops and duplicate edges... 58.696%   Checking for loops and duplicate edges... 60.870%   Checking for loops and duplicate edges... 63.043%   Checking for loops and duplicate edges... 65.217%   Checking for loops and duplicate edges... 67.391%   Checking for loops and duplicate edges... 69.565%   Checking for loops and duplicate edges... 71.739%   Checking for loops and duplicate edges... 73.913%   Checking for loops and duplicate edges... 76.087%   Checking for loops and duplicate edges... 78.261%   Checking for loops and duplicate edges... 80.435%   Checking for loops and duplicate edges... 82.609%   Checking for loops and duplicate edges... 84.783%   Checking for loops and duplicate edges... 86.957%   Checking for loops and duplicate edges... 89.130%   Checking for loops and duplicate edges... 91.304%   Checking for loops and duplicate edges... 93.478%   Checking for loops and duplicate edges... 95.652%   Checking for loops and duplicate edges... 97.826%   Checking for loops and duplicate edges... 100.000%
#>    Found and removed 4 loop(s)
#>    Found and removed 2 duplicate edge(s)